Removed unnecessary HelloWorld files. Reorganized files. Renamed folder. Removed unused css file. Code cleanup
