--- HelloWorld.java
@@ -1,8 +1,0 @@
-//App that prints "Hello World".
-//Will be used to test applet in html
-
-public class HelloWorld{
-	public static void main(String[] args){
-		System.out.println("Hello World");
-	}
-}
