+++ src/main/java/Analysis/FaceRectangle.java
@@ -1,4 +1,4 @@
-package Analysis;
+package analysis;
 
 public class FaceRectangle {
