+++ src/main/java/Analysis/FacialAnalysis.java
@@ -1,4 +1,4 @@
-package Analysis;// // This sample uses the Apache HTTP client from HTTP Components (http://hc.apache.org/httpcomponents-client-ga/)
+package analysis;// // This sample uses the Apache HTTP client from HTTP Components (http://hc.apache.org/httpcomponents-client-ga/)
 
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;
