+++ src/main/java/controller/Application.java
@@ -1,4 +1,4 @@
-package hello;
+package controller;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
