--- src/main/java/mystyle.css
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-
-<head>
- <style>
- header{
- 	text-align: center;}
- 
-
- html,body{
- 	 	background-image: url(background.JPG);}
-
-</style>
-</head>
