+++ src/main/resources/templates/analysis.html
@@ -1,8 +1,7 @@
 <!DOCTYPE HTML>
 <html xmlns:th="http://www.thymeleaf.org">
 <head>
-    <title>Getting Started: Serving Web Content</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+    <title> Halloween Spectacular: Which Superhero Are YOu?</title>
 </head>
 <body>
 <p th:text="'Hello, ' + ${url} + '!'" />
